fix(client-android): file notification — long to int for SignalR On
SignalR connection.On with long type — JSON number deserializes
as int by default. Type mismatch caused SignalR to silently drop
the ReceiveFileTransfer message — handler never called.

Changed long to int in ISignalService event, SignalService On
registration, App.xaml.cs handler.

+++ src/PcBeaconAgent.Client.Android/App.xaml.cs
@@ -167,7 +167,7 @@ public partial class App : Application
         }
     }
 
-    private void OnFileTransferReceived(string sourceIp, string fileName, long sizeBytes, string downloadUrl, string sourceMachine)
+    private void OnFileTransferReceived(string sourceIp, string fileName, int sizeBytes, string downloadUrl, string sourceMachine)
     {
         mLogger.LogInformation("File transfer received from {Ip} ({Machine}): {File} ({Size} bytes)",
             sourceIp, sourceMachine, fileName, sizeBytes);
